requirement.txt
discord==1.0.1
discord.py==1.6.0
python-dotenv==0.15.0
bs4
openai
requests
lxml